Posts file and folder restructure
