+++ apps/posts/src/App.tsx
@@ -1,4 +1,4 @@
-import PostAnalytics from './pages/PostAnalytics';
+import PostAnalytics from './views/post-analytics/PostAnalytics';
 import {FrameworkProvider, TopLevelFrameworkProps} from '@tryghost/admin-x-framework';
 import {RoutingProvider} from '@tryghost/admin-x-framework/routing';
 import {ShadeApp, ShadeAppProps} from '@tryghost/shade';
