+++ apps/posts/src/views/post-analytics/components/overview/NewsletterPerformance.tsx
